Grosse update (préparation lot 4)
- Modif chute pierre, pilier
- Ajout iOS, HTML5, Android
- Et pleins de corrections
- Bug a corriger : mineur bloqué à la génération de la base

+++ core/src/com/mygdx/mehelpers/Deplacement/Collision.java
@@ -26,7 +26,7 @@ public class Collision {
     };
     private Rectangle mineurRect;
     private int debutX, debutY, finX, finY;
-    private final float largeurMap;
+    private final int largeurMap;
     
     /**
      * @param deplacement objet Deplacement
